Refactor fetch requests. Rename fetch.js to api.js
Moved all fetch requests to api.js file.
